# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/add": {
            "post": {
                "description": "adds todolist item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoItem"
                ],
                "summary": "Add todo item",
                "parameters": [
                    {
                        "description": "todo item for add",
                        "name": "todoItem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.addTodoItemDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/get": {
            "get": {
                "description": "gets todolist item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoItem"
                ],
                "summary": "Get todo item by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id for search todo item",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TodoItem"
                        }
                    }
                }
            }
        },
        "/update": {
            "put": {
                "description": "updates todolist item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoItem"
                ],
                "summary": "Update todo item",
                "parameters": [
                    {
                        "description": "todo item for update",
                        "name": "todoItem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TodoItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "main.addTodoItemDto": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.TodoItem": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}